requirements.txt
requests>=2.23.0
PyYAML>=5.3
